+++ tests/test_config/test_blank_line_after_tag.py
@@ -213,8 +213,11 @@ test_data = [
         ),
         (
             '<svg viewBox="0 0 24 24">\n'
-            '    <path stroke-linecap="round" stroke-linejoin="round" '
-            'd=" {% block svg_path %} {% endblock svg_path %}  " />\n'
+            '    <path stroke-linecap="round" stroke-linejoin="round" d="\n'
+            "  {% block svg_path %}\n"
+            "  {% endblock svg_path %}\n"
+            "\n"
+            '  " />\n'
             "</svg>\n"
         ),
         ({"blank_line_after_tag": "endblock"}),